# docs/api-ledger/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "CreateCustomerResource" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "contactEmail" : {
            "format" : "email",
            "type" : "string"
          },
          "legacyCompanyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "ruc" : {
            "maxLength" : 11,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
...
      "CreateSiteResource" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "name" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "CustomerResource" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "address" : {
            "type" : "string"
          },
          "contactEmail" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "legacyCompanyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "organizationId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "phone" : {
            "type" : "string"
          },
          "ruc" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SiteResource" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "address" : {
            "type" : "string"
          },
          "customerAccountId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "organizationId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token JWT Bearer para autenticar solicitudes a la API",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "support@primefuel.com",
      "name" : "Soporte de FullTank",
      "url" : "https://primefuel.com/support"
    },
    "description" : "API de FullTank para gestionar abastecimiento, flota, entregas, inventario y pagos por organizaciÃ³n.",
    "license" : {
      "name" : "Apache 2.0 (licencia)",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "FullTank Platform",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v2/customers" : {
      "get" : {
        "description" : "Devuelve las cuentas asociadas a la organización activa del usuario.",
        "operationId" : "listCustomers",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CustomerResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Cuentas de cliente devueltas."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CustomerResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "El usuario no está autenticado o no tiene una organización activa."
          }
        },
        "summary" : "Listar cuentas de cliente",
        "tags" : [ "Clientes" ]
      },
      "post" : {
        "description" : "Crea la cuenta dentro de la organización activa del usuario. La organización se deriva de la membresía autenticada.",
        "operationId" : "registerCustomer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCustomerResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Cuenta de cliente creada."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El cuerpo no cumple las validaciones requeridas."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El usuario no está autenticado o no tiene una organización activa."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Ya existe una cuenta con el mismo RUC en esta organización."
          }
        },
        "summary" : "Registrar una cuenta de cliente",
        "tags" : [ "Clientes" ]
      }
    },
    "/api/v2/customers/{customerId}/sites" : {
      "get" : {
        "description" : "Devuelve los sitios registrados bajo una cuenta que pertenece a la organización activa.",
        "operationId" : "listSites",
        "parameters" : [ {
          "in" : "path",
          "name" : "customerId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SiteResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Sitios de entrega devueltos."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SiteResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "El usuario no está autenticado o no tiene una organización activa."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SiteResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "La cuenta no existe o pertenece a otra organización."
          }
        },
        "summary" : "Listar sitios de una cuenta",
        "tags" : [ "Clientes" ]
      },
      "post" : {
        "description" : "Crea un sitio bajo la cuenta indicada, que debe pertenecer a la organización activa del usuario.",
        "operationId" : "registerSite",
        "parameters" : [ {
          "in" : "path",
          "name" : "customerId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSiteResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Sitio creado."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El cuerpo no cumple las validaciones requeridas."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El usuario no está autenticado, no tiene organización activa o la cuenta pertenece a otro tenant."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "No existe la cuenta de cliente indicada."
          }
        },
        "summary" : "Registrar un sitio de entrega",
        "tags" : [ "Clientes" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Servidor de la API",
    "url" : "http://localhost:8080"
  } ],
...
}